+++ src/retentioneering/data_processors/__init__.py
@@ -8,6 +8,7 @@ from retentioneering.data_processors.edit_events import EditEvents
 from retentioneering.data_processors.filter_events import FilterEvents
 from retentioneering.data_processors.filter_paths import FilterPaths
 from retentioneering.data_processors.rename_events import RenameEvents
+from retentioneering.data_processors.rename_segment_values import RenameSegmentValues
 from retentioneering.data_processors.sample_paths import SamplePaths
 from retentioneering.data_processors.split_sessions import SplitSessions
 from retentioneering.data_processors.truncate_paths import TruncatePaths
@@ -25,6 +26,7 @@ __all__ = [
     "FilterEvents",
     "FilterPaths",
     "RenameEvents",
+    "RenameSegmentValues",
     "SamplePaths",
     "SplitSessions",
     "TruncatePaths",